New menu
so everything should have the new menu that i added. One main difference is when you look through the html for the /games they are slightly different when it comes to using href.  I had to use ../ anything that wasnt located in the games file but it should all work. The ones located outside the games file should be fine as well just small changes like how if you check i used games/Clicker.html. The css is kept the same throughout the entire css in the ones i implemented it. Let me know if it works!