Changes to mempool only
this fixes the sequential part of the mempool. Only file changed was the mempool and the css. The css still needs to me fix so that you can scroll normally. The video i am gonna send should explain what im talking about. I only fixed the mempool but not the block chain so youll need to fix the block chain java script file and you will still need to make it 5 transactions. Ardion you want all you have to do is just fix the blockchain and ill do the rest when im free